api: Correct handling of children in some text replacements
Fixes GH-232
