+++ api/src/main/java/net/kyori/adventure/text/TextReplacementRenderer.java
@@ -77,6 +77,9 @@ final class TextReplacementRenderer implements ComponentRenderer<TextReplacement
               .style(component.style()));
 
             modified = replacement == null ? Component.empty() : replacement.asComponent();
+            if(children == null) { // Prepare children
+              children = new ArrayList<>(oldChildrenSize);
+            }
           } else {
             // otherwise, work on a child of the root node
             modified = Component.text("", component.style());
